update valueType support
