+++ packages/zent-datetimepicker/README.md
@@ -214,7 +214,7 @@ ReactDOM.render(
 | format       | 返回日期字符串格式                | string         | `YYYY-MM-DD`  | 否    |
 | min        | 可选日期的最小值                   | string/Date    |     | 否    |
 | max        | 可选日期的最大值                   | string/Date    |     | 否    |
-| returnType | 设置 onChange 的返回值，可选值为 `string`/`number`/`date`  | string     | '' | 否    |
+| valueType | 设置 onChange 的返回值，可选值为 `string`/`number`/`date`  | string     | '' | 否    |
 | placeholder  | 提示文案                   | string    | `请选择日期`   | 否    |
 
 
@@ -247,7 +247,7 @@ ReactDOM.render(
 | format       | 返回日期字符串格式  | string | `'YYYY-MM-DD'` | 否    |
 | min          | 可选日期的最小值   | string/instanceOf(Date)  | ``   | 否    |
 | max          | 可选日期的最大值  | string/instanceOf(Date)  | ``    | 否    |
-| returnType | 设置 onChange 的返回值，可选值为 `string`/`number`/`date`  | string     | '' | 否    |
+| valueType | 设置 onChange 的返回值，可选值为 `string`/`number`/`date`  | string     | '' | 否    |
 | placeholder  | 提示文案          | string    | `['开始日期','结束日期']`   | 否    |
