+++ packages/zent-datetimepicker/examples/datePicker.js
@@ -68,17 +68,18 @@ export default class Simple extends Component {
           format="YYYY年MM月DD日"
           min="2017/01/01"
           max="2017/11/11"
+          valueType="number"
           disabledTime={this.isDisabledTime}
           onChange={this.onChangeDate}
           value={state.value}
         />
-        <button type="button" onClick={this.reset}>reset</button>
         <br /><br />
         <DatePicker
           showTime
           className="top-right"
           format="YYYY-MM-DD HH:mm:ss"
           placeholder="请选择日期和时间"
+          valueType="number"
           disabledDate={this.isDisabledDate}
           disabledTime={this.isDisabledTime}
           onChange={this.onChangeDate}
@@ -89,6 +90,7 @@ export default class Simple extends Component {
           className="bottom-right"
           format="YYYY-MM-DD HH:mm:ss"
           placeholder="请选择日期和时间"
+          valueType="number"
           disabledDate={this.isDisabledDate}
           disabledTime={this.isDisabledTime}
           onChange={this.onChangeDate}
