+++ packages/zent-datetimepicker/src/utils/index.js
@@ -1,5 +1,3 @@
-import Popover from 'zent-popover';
-
 export const CURRENT = new Date();
 export const CURRENT_DAY = new Date(CURRENT.getFullYear(), CURRENT.getMonth(), CURRENT.getDate());
 export const CURRENT_YEAR = CURRENT.getFullYear();
@@ -67,33 +65,3 @@ export const isFunction = (val) => {
   return Object.prototype.toString.call(val) === '[object Function]';
 };
 
-export const positionMap = {
-  'bottom-left': Popover.Position.BottomLeft,
-  'bottom-right': Popover.Position.BottomRight,
-  'top-left': Popover.Position.TopLeft,
-  'top-right': Popover.Position.TopRight
-};
-
-export const position = Popover.Position.create((anchorBoundingBox, containerBoundingBox, contentDimension, options) => {
-  const windowWidth = window.innerWidth;
-  const windowHeight = window.innerHeight;
-  const { anchorBoundingBoxViewport, cushion } = options;
-
-  let horizontal;
-  let vertical;
-  if (anchorBoundingBoxViewport.left + contentDimension.width > windowWidth) {
-    horizontal = 'right';
-  } else {
-    horizontal = 'left';
-  }
-
-  if (anchorBoundingBoxViewport.top + anchorBoundingBoxViewport.height + cushion + contentDimension.height > windowHeight) {
-    vertical = 'top';
-  } else {
-    vertical = 'bottom';
-  }
-
-  const key = `${vertical}-${horizontal}`;
-
-  return positionMap[key]('zent', anchorBoundingBox, containerBoundingBox, contentDimension, options);
-});
